// src/types/api.ts
export interface Bucket {
	name: string;
	ownerId: string;
	createdAt: string; // ISO 8601
	objectCount?: number;
	totalObjectBytes?: number;
}